fix: move mixin to boot file, update store imports
