+++ quasar.conf.js
@@ -22,7 +22,7 @@ module.exports = function(/* ctx */) {
     // app boot file (/src/boot)
     // --> boot files are part of "main.js"
     // https://quasar.dev/quasar-cli/boot-files
-    boot: ["ual", "hyperion", "api", "errorHandling", "helpers"],
+    boot: ["ual", "hyperion", "api", "errorHandling", "helpers", "mixin"],
 
     // https://quasar.dev/quasar-cli/quasar-conf-js#Property%3A-css
     css: ["app.scss"],
